remove CNY from surrencyArray
since CNY is not recognized by the coinAPI then I need to delete it from the currencyArray to avoid error'

This update still need to be tested, thus it need to be put to test by local XCode for testing. fetch it, check it out first and then test it.